feat: 프로필 이미지가 없는 유저의 delete 요청 처리

+++ src/module/user/adapter/presenter/rest/user.controller.ts
@@ -88,6 +88,10 @@ export class UserController {
   })
   @Delete('profile/image')
   async deleteUserProfileImage(@ActiveUser() user: UserEntity) {
+    if (!user.profile?.profileImageUrl) {
+      return user;
+    }
+
     return this.userService.deleteProfileImage({
       id: user.id,
       profileImageUrl: user.profile!.profileImageUrl!,
